requirements.txt
django==5.1
djangorestframework==pip3.15.2
celery==5.4.0
redis==5.0.8
django-celery-beat==2.7.0
black==24.8.0